requirements.txt
attr==0.3.1
attrs==21.2.0
Django==4.0.3
djangorestframework==3.13.1
drf_yasg==1.20.0
celery==5.2.6
